Rename filesystem files to match their namespace
