+++ mk/newsboat.deps
@@ -1,1 +1,1 @@
-newsboat.cpp src/cache.cpp  src/htmlrenderer.cpp src/urlreader.cpp src/view.cpp src/controller.cpp src/reloadthread.cpp src/tagsouppullparser.cpp src/downloadthread.cpp src/rssignores.cpp src/rssparser.cpp src/formaction.cpp src/listformaction.cpp src/feedlistformaction.cpp src/itemlistformaction.cpp src/itemviewformaction.cpp src/helpformaction.cpp src/dirbrowserformaction.cpp src/filebrowserformaction.cpp src/urlviewformaction.cpp src/selectformaction.cpp src/history.cpp src/filtercontainer.cpp src/listformatter.cpp src/regexmanager.cpp src/dialogsformaction.cpp src/ttrssapi.cpp src/ttrssurlreader.cpp src/newsblurapi.cpp src/newsblururlreader.cpp src/oldreaderurlreader.cpp src/oldreaderapi.cpp src/feedcontainer.cpp src/feedhqapi.cpp src/feedhqurlreader.cpp src/freshrssapi.cpp src/freshrssurlreader.cpp src/textformatter.cpp src/ocnewsapi.cpp src/ocnewsurlreader.cpp src/remoteapi.cpp src/inoreaderapi.cpp src/inoreaderurlreader.cpp src/cliargsparser.cpp src/configpaths.cpp src/reloader.cpp src/reloadrangethread.cpp src/opml.cpp src/fileurlreader.cpp src/opmlurlreader.cpp src/itemrenderer.cpp src/queuemanager.cpp src/rssitem.cpp src/rssfeed.cpp src/listwidget.cpp src/textviewwidget.cpp src/regexowner.cpp src/configactionhandler.cpp src/minifluxapi.cpp src/minifluxurlreader.cpp src/emptyformaction.cpp src/statusline.cpp src/filesystembrowser.cpp src/searchresultslistformaction.cpp src/itemutils.cpp
+newsboat.cpp src/cache.cpp  src/htmlrenderer.cpp src/urlreader.cpp src/view.cpp src/controller.cpp src/reloadthread.cpp src/tagsouppullparser.cpp src/downloadthread.cpp src/rssignores.cpp src/rssparser.cpp src/formaction.cpp src/listformaction.cpp src/feedlistformaction.cpp src/itemlistformaction.cpp src/itemviewformaction.cpp src/helpformaction.cpp src/dirbrowserformaction.cpp src/filebrowserformaction.cpp src/urlviewformaction.cpp src/selectformaction.cpp src/history.cpp src/filtercontainer.cpp src/listformatter.cpp src/regexmanager.cpp src/dialogsformaction.cpp src/ttrssapi.cpp src/ttrssurlreader.cpp src/newsblurapi.cpp src/newsblururlreader.cpp src/oldreaderurlreader.cpp src/oldreaderapi.cpp src/feedcontainer.cpp src/feedhqapi.cpp src/feedhqurlreader.cpp src/freshrssapi.cpp src/freshrssurlreader.cpp src/textformatter.cpp src/ocnewsapi.cpp src/ocnewsurlreader.cpp src/remoteapi.cpp src/inoreaderapi.cpp src/inoreaderurlreader.cpp src/cliargsparser.cpp src/configpaths.cpp src/reloader.cpp src/reloadrangethread.cpp src/opml.cpp src/fileurlreader.cpp src/opmlurlreader.cpp src/itemrenderer.cpp src/queuemanager.cpp src/rssitem.cpp src/rssfeed.cpp src/listwidget.cpp src/textviewwidget.cpp src/regexowner.cpp src/configactionhandler.cpp src/minifluxapi.cpp src/minifluxurlreader.cpp src/emptyformaction.cpp src/statusline.cpp src/file_system.cpp src/searchresultslistformaction.cpp src/itemutils.cpp
